+++ src/main/java/xd/arkosammy/creeperhealing/blocks/DoubleAffectedBlock.java
@@ -11,7 +11,7 @@ import net.minecraft.state.property.Properties;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
 import net.minecraft.world.World;
-import xd.arkosammy.creeperhealing.configuration.ReplaceMapConfig;
+import xd.arkosammy.creeperhealing.config.ReplaceMapConfig;
 import xd.arkosammy.creeperhealing.explosions.AbstractExplosionEvent;
 import xd.arkosammy.creeperhealing.util.ExplosionUtils;
