+++ src/main/java/xd/arkosammy/creeperhealing/commands/ExplosionSourcesCommands.java
@@ -8,7 +8,7 @@ import com.mojang.brigadier.tree.LiteralCommandNode;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.text.Text;
-import xd.arkosammy.creeperhealing.configuration.ExplosionSourceConfig;
+import xd.arkosammy.creeperhealing.config.ExplosionSourceConfig;
 
 public final class ExplosionSourcesCommands {
