+++ src/main/java/xd/arkosammy/creeperhealing/commands/HealingCommandManager.java
@@ -9,7 +9,7 @@ import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
-import xd.arkosammy.creeperhealing.configuration.CreeperHealingConfig;
+import xd.arkosammy.creeperhealing.config.CreeperHealingConfig;
 
 import java.io.IOException;
