+++ src/main/java/xd/arkosammy/creeperhealing/commands/ModeCommands.java
@@ -6,7 +6,7 @@ import com.mojang.brigadier.tree.LiteralCommandNode;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.text.Text;
-import xd.arkosammy.creeperhealing.configuration.ModeConfig;
+import xd.arkosammy.creeperhealing.config.ModeConfig;
 import xd.arkosammy.creeperhealing.explosions.ExplosionHealingMode;
 
 public final class ModeCommands {
