+++ src/main/java/xd/arkosammy/creeperhealing/mixin/BlockMixin.java
@@ -15,7 +15,7 @@ import net.minecraft.entity.vehicle.TntMinecartEntity;
 import net.minecraft.world.explosion.Explosion;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import xd.arkosammy.creeperhealing.configuration.ExplosionItemDropConfig;
+import xd.arkosammy.creeperhealing.config.ExplosionItemDropConfig;
 import xd.arkosammy.creeperhealing.util.ExplosionUtils;
 import xd.arkosammy.creeperhealing.explosions.ducks.ExplosionAccessor;
